Update cast conversion to handle extreme values

+++ functions.c
@@ -113,8 +113,8 @@ int print_int(va_list types, char buffer[],
 {
 	int i = BUFF_SIZE - 2;
 	int is_negative = 0;
-	long int n = va_arg(types, long int);
-	unsigned int num;
+	long long n = va_arg(types, long long);
+	unsigned long long num;
 
 	n = convert_size_number(n, size);
 
@@ -122,11 +122,11 @@ int print_int(va_list types, char buffer[],
 		buffer[i--] = '0';
 
 	buffer[BUFF_SIZE - 1] = '\0';
-	num = (unsigned int)n;
+	num = (unsigned long long)n;
 
 	if (n < 0)
 	{
-		num = (unsigned int)((-1) * n);
+		num = (unsigned long long)((-1) * n);
 		is_negative = 1;
 	}
